wip: fixing connentions with users and leader messages between them

+++ common/src/constants.rs
@@ -1,6 +1,6 @@
 pub const COORDINATE_SCALE: f32 = 10.0;
 pub const SUCCESS_PROBABILITY: f32 = 0.75;
-pub const NUM_COORDINATORS: usize = 4;
+pub const NUM_COORDINATORS: u16 = 4;
 pub const BASE_PORT: u16 = 8080;
 pub const TIMEOUT_SECONDS: u64 = 2;
 const DELAY_SECONDS: u64 = 2;
